feat: separating independent services
