+++ oracle-api/src/main/java/ca/bc/gov/oracleapi/service/consep/MoistureContentService.java
@@ -1,7 +1,6 @@
 package ca.bc.gov.oracleapi.service.consep;
 
 import ca.bc.gov.oracleapi.config.SparLog;
-import ca.bc.gov.oracleapi.dto.consep.ActivityFormDto;
 import ca.bc.gov.oracleapi.dto.consep.MoistureContentConesDto;
 import ca.bc.gov.oracleapi.dto.consep.MccReplicateDto;
 import ca.bc.gov.oracleapi.dto.consep.MccReplicateFormDto;
@@ -15,7 +14,6 @@ import ca.bc.gov.oracleapi.repository.consep.MccReplicatesRepository;
 import ca.bc.gov.oracleapi.repository.consep.TestResultRepository;
 import jakarta.transaction.Transactional;
 import java.math.BigDecimal;
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -152,39 +150,6 @@ public class MoistureContentService {
   }
 
   /**
-   * Update activity table.
-   *
-   * @param riaKey the identifier key for all table related to MCC
-   * @param activityFormDto an object with the values to be updated
-   */
-  @Transactional
-  public ActivityEntity updateActivityField(
-      @NonNull BigDecimal riaKey,
-      @NonNull ActivityFormDto activityFormDto
-  ) {
-    SparLog.info("Updating activity with riaKey: {}", riaKey);
-
-    ActivityEntity activity = activityRepository.findById(riaKey)
-        .orElseGet(() -> {
-          SparLog.info("No existing activity found for riaKey: {}. Creating a new one.", riaKey);
-          ActivityEntity newActivity = new ActivityEntity();
-          newActivity.setRiaKey(riaKey);
-          return newActivity;
-        });
-
-    activity.setActualBeginDateTime(activityFormDto.actualBeginDateTime());
-    activity.setActualEndDateTime(activityFormDto.actualEndDateTime());
-    activity.setTestCategoryCode(activityFormDto.testCategoryCode());
-    activity.setRiaComment(activityFormDto.riaComment());
-
-    ActivityEntity savedActivity = activityRepository.save(activity);
-
-    SparLog.info("Activity with riaKey: {} saved successfully.", riaKey);
-
-    return savedActivity;
-  }
-
-  /**
    * This function validates the data to be submitted for MCC.
    * Throws exception if validation fails
    *
@@ -248,75 +213,6 @@ public class MoistureContentService {
           "Container and dry weight is missing or invalid");
       }
     }
-  }
-
-  /**
-   * This function validates Activity part of the data to be submitted for MCC.
-   * Throws exception if validation fails.
-   *
-   * @param activityData activity entity to be validated
-   * @throws ResponseStatusException if any validation fails
-   */
-  public void validateMoistureContentActivityData(ActivityEntity activityData) {
-    SparLog.info("Validating MCC activity data");
-
-    if (activityData.getTestCategoryCode() == null) {
-      SparLog.error("MCC activity data validation failed: Test category code is missing");
-      throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Test category code is missing");
-    }
-    if (activityData.getActualBeginDateTime() == null
-        || activityData.getActualBeginDateTime().compareTo(LocalDateTime.now()) < 0) {
-      SparLog.error("MCC activity data validation failed: "
-          + "Actual begin date time is missing or in the past");
-      throw new ResponseStatusException(
-        HttpStatus.BAD_REQUEST,
-        "Actual begin date time is missing or in the past");
-    }
-    if (activityData.getActualEndDateTime() == null
-        || activityData.getActualEndDateTime().compareTo(LocalDateTime.now()) < 0) {
-      SparLog.error("MCC activity data validation failed: "
-          + "Actual end date time is missing or in the past");
-      throw new ResponseStatusException(
-        HttpStatus.BAD_REQUEST,
-        "Actual end date time is missing or in the past");
-    }
-  }
-
-  /**
-   * Update the test status to "completed".
-   *
-   * @param riaKey the identifier key for all table related to MCC
-   */
-  public void updateTestResultStatusToCompleted(BigDecimal riaKey) {
-    SparLog.info("Updating test result status to completed for RIA_SKEY: {}", riaKey);
-
-    testResultRepository.updateTestResultStatusToCompleted(riaKey);
-
-    SparLog.info("Test result status updated to completed for RIA_SKEY: {}", riaKey);
-  }
-
-  /**
-   * Accept the test results for the given riaKey.
-   *
-   * @param riaKey the identifier key for all table related to MCC
-   */
-  public void acceptMoistureContentData(BigDecimal riaKey) {
-    SparLog.info("Accepting moisture content data for RIA_SKEY: {}", riaKey);
-
-    Optional<MoistureContentConesDto> moistureContent = getMoistureConeContentData(riaKey);
-
-    if (moistureContent.isEmpty()) {
-      SparLog.warn("No data found for RIA_SKEY: {}", riaKey);
-      throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No data found for given RIA_KEY");
-    }
-
-    if (moistureContent.get().testCompleteInd() == 0) {
-      SparLog.error("Test is not completed for RIA_SKEY: {}", riaKey);
-      throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Test is not completed");
-    }
-
-    testResultRepository.updateTestResultStatusToAccepted(riaKey);
-    SparLog.info("Moisture content data accepted for RIA_SKEY: {}", riaKey);
   }
 
   /**
